AWML_dev_tools/compare_video_result/pyproject.toml
[project]
name = "compare-video-result"
version = "0.1.0"
description = "A video concatenation tool to compare results from ml models."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moviepy",
    "PyYAML",
]
